Update code to use LondonTireChangeTime instead of TireChangeTime
This commit changes the use of the TireChangeTime model to LondonTireChangeTime across various classes. The update includes modifying method return types, arguments, and variable types. These changes seek to fix 500 error that occured while parsing XML and reason for this was using wrong model, that did not had JacksonXML annotation.