Refactor service methods to accept lists of parameters
Changed the string parameter of the methods findAvailableTimes in TireShopService and WebController to accept a list of strings. This allows multiple tire shops and vehicle types to be processed simultaneously, improving the flexibility of the service. Additionally, removed a redundant findAvailableTimes method in MainController and updated corresponding JavaScript and HTML code to match these changes.