Modify method access modifier and add related test
Changed method 'mapManchesterToTireChangeBooking' from private to the default in 'TireShopService.java'. Also, created a new file 'TireShopServiceTest.java' to add a test case ensuring that fields are mapped correctly in this method. This test will assist in verifying the logic of the method and help catch any potential bugs or regression in the future.